+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.3 (WIP)
 
 * Fix `al-file` element showing wrong buttons on load
+* Always add `role="presentation"` to images that get a simple placeholder
 
 ## 0.6.2 (2022-07-06)